# expression: implement vectorized evaluation for `builtinArithmeticMinusIntSig`