--- a/cmd/explaintest/r/topn_push_down.result
+++ b/cmd/explaintest/r/topn_push_down.result
@@ -218,8 +218,8 @@ explain select /*+ TIDB_INLJ(t2) */ * from t t1 join t t2 on t1.a = t2.a limit 5
 id	count	task	operator info
 Limit_11	5.00	root	offset:0, count:5
 └─IndexMergeJoin_19	5.00	root	inner join, inner:IndexReader_17, outer key:Column#1, inner key:Column#3
-  ├─IndexReader_25	4.00	root	index:IndexScan_24
-  │ └─IndexScan_24	4.00	cop[tikv]	table:t1, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
+  ├─TableReader_23	4.00	root	data:TableScan_22
+  │ └─TableScan_22	4.00	cop[tikv]	table:t1, range:[-inf,+inf], keep order:false, stats:pseudo
   └─IndexReader_17	1.25	root	index:IndexScan_16
     └─IndexScan_16	1.25	cop[tikv]	table:t2, index:a, range: decided by [eq(Column#3, Column#1)], keep order:true, stats:pseudo
 explain select /*+ TIDB_INLJ(t2) */ * from t t1 left join t t2 on t1.a = t2.a where t2.a is null limit 5;
@@ -227,8 +227,8 @@ id	count	task	operator info
 Limit_12	5.00	root	offset:0, count:5
 └─Selection_13	5.00	root	isnull(Column#3)
   └─IndexMergeJoin_21	5.00	root	left outer join, inner:IndexReader_19, outer key:Column#1, inner key:Column#3
-    ├─IndexReader_27	4.00	root	index:IndexScan_26
-    │ └─IndexScan_26	4.00	cop[tikv]	table:t1, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
+    ├─TableReader_25	4.00	root	data:TableScan_24
+    │ └─TableScan_24	4.00	cop[tikv]	table:t1, range:[-inf,+inf], keep order:false, stats:pseudo
     └─IndexReader_19	1.25	root	index:IndexScan_18
       └─IndexScan_18	1.25	cop[tikv]	table:t2, index:a, range: decided by [eq(Column#3, Column#1)], keep order:true, stats:pseudo
 explain select /*+ TIDB_SMJ(t1, t2) */ * from t t1 join t t2 on t1.a = t2.a limit 5;
@@ -251,17 +251,17 @@ Limit_12	5.00	root	offset:0, count:5
 explain select /*+ TIDB_HJ(t1, t2) */ * from t t1 join t t2 on t1.a = t2.a limit 5;
 id	count	task	operator info
 Limit_11	5.00	root	offset:0, count:5
-└─HashLeftJoin_31	5.00	root	inner join, inner:IndexReader_40, equal:[eq(Column#1, Column#3)]
-  ├─IndexReader_36	4.00	root	index:IndexScan_35
-  │ └─IndexScan_35	4.00	cop[tikv]	table:t1, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
-  └─IndexReader_40	10000.00	root	index:IndexScan_39
-    └─IndexScan_39	10000.00	cop[tikv]	table:t2, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
+└─HashLeftJoin_31	5.00	root	inner join, inner:TableReader_38, equal:[eq(Column#1, Column#3)]
+  ├─TableReader_34	4.00	root	data:TableScan_33
+  │ └─TableScan_33	4.00	cop[tikv]	table:t1, range:[-inf,+inf], keep order:false, stats:pseudo
+  └─TableReader_38	10000.00	root	data:TableScan_37
+    └─TableScan_37	10000.00	cop[tikv]	table:t2, range:[-inf,+inf], keep order:false, stats:pseudo
 explain select /*+ TIDB_HJ(t1, t2) */ * from t t1 left join t t2 on t1.a = t2.a where t2.a is null limit 5;
 id	count	task	operator info
 Limit_12	5.00	root	offset:0, count:5
 └─Selection_13	5.00	root	isnull(Column#3)
-  └─HashLeftJoin_24	5.00	root	left outer join, inner:IndexReader_32, equal:[eq(Column#1, Column#3)]
-    ├─IndexReader_28	4.00	root	index:IndexScan_27
-    │ └─IndexScan_27	4.00	cop[tikv]	table:t1, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
-    └─IndexReader_32	10000.00	root	index:IndexScan_31
-      └─IndexScan_31	10000.00	cop[tikv]	table:t2, index:a, range:[NULL,+inf], keep order:false, stats:pseudo
+  └─HashLeftJoin_24	5.00	root	left outer join, inner:TableReader_30, equal:[eq(Column#1, Column#3)]
+    ├─TableReader_26	4.00	root	data:TableScan_25
+    │ └─TableScan_25	4.00	cop[tikv]	table:t1, range:[-inf,+inf], keep order:false, stats:pseudo
+    └─TableReader_30	10000.00	root	data:TableScan_29
+      └─TableScan_29	10000.00	cop[tikv]	table:t2, range:[-inf,+inf], keep order:false, stats:pseudo