--- a/config/config.toml.example
+++ b/config/config.toml.example
@@ -319,8 +319,8 @@ enable = true
 max-retry-count = 256
 
 # default TTL in milliseconds for pessimistic lock.
-# The value must between "15s" and "120s".
-ttl = "40s"
+# The value must between "3s" and "30s".
+ttl = "10s"
 
 [stmt-summary]
 # max number of statements kept in memory.