--- a/ddl/ddl.go
+++ b/ddl/ddl.go
@@ -256,6 +256,7 @@ type DDL interface {
 	LockTables(ctx sessionctx.Context, stmt *ast.LockTablesStmt) error
 	UnlockTables(ctx sessionctx.Context, lockedTables []model.TableLockTpInfo) error
 	CleanupTableLock(ctx sessionctx.Context, tables []*ast.TableName) error
+	UpdateTableReplicaInfo(ctx sessionctx.Context, tid int64, available bool) error
 
 	// GetLease returns current schema lease time.
 	GetLease() time.Duration