--- a/executor/show.go
+++ b/executor/show.go
@@ -1098,6 +1098,9 @@ func (e *ShowExec) fetchShowPumpOrDrainerStatus(kind string) error {
 	}
 
 	for _, n := range nodes {
+		if n.State == node.Offline {
+			continue
+		}
 		e.appendRow([]interface{}{n.NodeID, n.Addr, n.State, n.MaxCommitTS, utils.TSOToRoughTime(n.UpdateTS).Format(types.TimeFormat)})
 	}