… config/config_test.go
… ddl/db_test.go
… executor/admin_test.go
… executor/builder.go
… executor/executor.go
… executor/statement_context_test.go
… expression/builtin_arithmetic_vec.go
… expression/builtin_arithmetic_vec_test.go
… expression/builtin_cast.go
… expression/builtin_cast_vec.go
… expression/builtin_compare_vec.go
… expression/builtin_compare_vec_generated.go
… expression/builtin_compare_vec_generated_test.go
… expression/builtin_compare_vec_test.go
… expression/builtin_json_vec.go
… expression/builtin_json_vec_test.go
… expression/builtin_time_vec.go
… expression/builtin_time_vec_test.go
… expression/column.go
… expression/column_test.go
… expression/constant_fold.go
… expression/expression.go
… expression/generator/compare_vec.go
… expression/integration_test.go
… expression/util.go
… expression/util_test.go
… go.mod
… go.sum
… infoschema/perfschema/const.go
… infoschema/perfschema/tables.go